Importer script
Started writing an importer script to pull data from the old database
and write it into an SQL file to import into the new database.

For multiple reasons I have decided to write this in Ruby. Yes it adds
another language to the project, but the Mongo bindings for Rust aren't
good so it was this or Python. Or like Bash...