Added external flag to projects.
I've added a external variable to the Project and NewProject structs. The database has been updated as well.

Currently, the flag can only be toggled in the Edit Project webpage and viewed in the Project webpage. I need to figure out how to add it to the New Project webpage as well.