Fix Dashboard and User pages
Fixes issue #74

When a new project is created, if the repo given was not a valid Git repo it was failing.

The funny thing is that I acutally did write graceful error handling for this.
And then just called `unwrap` in it...

A simple switch from a `map` to a `filter_map` fixes it all.
I did some extra error messages too while I was at it.