Improve password hashing API
I had been unhappy with the way that we interacted
with the password hashing utilities for a while now

This commit removes gen_salt() and merges it into hash_password()
and introduces the HashedPassword struct
which is used to represent a hashed password in a safer manner.
I also added a simple type def UnsafeBinaryString
to make it more obvious that the output of a password hashing
is not a valid string but we are treating it like one.