Make UnsafeBinaryString a newtype struct
I remove the HashedPassword struct and replaced it with a more
robust UnsafeBinaryString struct that implements a derive
for all the Diesel types it needs to be compatible with the database.
With that all the calls to both() have been removed.

This should offer much more safety and generally be clearer.