Reserved words and expanded FormError use
Added a new type of FormError: ReservedWord
and a helper function for it. It checks against a list of words names
that are not allowed for projects/users/groups/etc. and raises a
FormError as needed.

With that I have expanded the use of FormError to as many forms as I can
so that the ReservedWord error and other errors can be presented to the user.