Date Formatting - Initial Implementation
I've implemented a cleaner date format to be used on the site (it took a while to find the proper NaiveDateTime string parses).

Anyway, I still need to see if I can implement this format functionality to the user and event pages. Those will be a little harder to implement. Other than that, I think everything has the nicer date format.