iCalendar support for the calendar
The calendar now supports automatically generating an iCalendar file
which can be subscribed to using a variety of services like Google
Calendar.

This required the `icalendar` crate and the creation of a new
`/calendar.ics` route to serve it.

Also made some fixes to the `calendar.js` file, notably fixing
formatting and line-endings.